+++ src/MFSL/template.CACHEINODE/cache_inode_link.c
@@ -119,13 +119,13 @@ fsal_status_t cache_inode_async_link(cache_inode_async_op_desc_t * popasyncdesc)
   fsal_status_t fsal_status;
 
   fsal_status = FSAL_link(popasyncdesc->op_args.link.pfsal_handle_src,
-			  popasyncdesc->op_args.link.pfsal_handle_dirdest,
-			  &popasyncdesc->op_args.link.name_link,
-			  &popasyncdesc->fsal_op_context,
-			  &popasyncdesc->op_res.link.attr);
+                          popasyncdesc->op_args.link.pfsal_handle_dirdest,
+                          &popasyncdesc->op_args.link.name_link,
+                          &popasyncdesc->fsal_op_context,
+                          &popasyncdesc->op_res.link.attr);
 
   return fsal_status;
-}				/* cache_inode_aync_rename_src */
+}                               /* cache_inode_aync_rename_src */
 
 /**
  *
@@ -149,13 +149,13 @@ fsal_status_t cache_inode_async_link(cache_inode_async_op_desc_t * popasyncdesc)
  *
  */
 cache_inode_status_t cache_inode_link(cache_entry_t * pentry_src,
-				      cache_entry_t * pentry_dir_dest,
-				      fsal_name_t * plink_name,
-				      fsal_attrib_list_t * pattr,
-				      hash_table_t * ht,
-				      cache_inode_client_t * pclient,
-				      fsal_op_context_t * pcontext,
-				      cache_inode_status_t * pstatus)
+                                      cache_entry_t * pentry_dir_dest,
+                                      fsal_name_t * plink_name,
+                                      fsal_attrib_list_t * pattr,
+                                      hash_table_t * ht,
+                                      cache_inode_client_t * pclient,
+                                      fsal_op_context_t * pcontext,
+                                      cache_inode_status_t * pstatus)
 {
   fsal_status_t fsal_status;
   fsal_handle_t handle_src;
@@ -195,9 +195,9 @@ cache_inode_status_t cache_inode_link(cache_entry_t * pentry_src,
 
   /* Check if an entry of the same name doesn't exist in the destination directory */
   if ((pentry_lookup = cache_inode_lookup(pentry_dir_dest,
-					  plink_name,
-					  &lookup_attributes,
-					  ht, pclient, pcontext, pstatus)) != NULL)
+                                          plink_name,
+                                          &lookup_attributes,
+                                          ht, pclient, pcontext, pstatus)) != NULL)
     {
       /* There exists such an entry... */
       *pstatus = CACHE_INODE_ENTRY_EXISTS;
@@ -258,8 +258,8 @@ cache_inode_status_t cache_inode_link(cache_entry_t * pentry_src,
 
     default:
       DisplayLogJdLevel(pclient->log_outputs, NIV_CRIT,
-			"WARNING: unknown source pentry type: internal_md.type=%d, line %d in file %s",
-			pentry_src->internal_md.type, __LINE__, __FILE__);
+                        "WARNING: unknown source pentry type: internal_md.type=%d, line %d in file %s",
+                        pentry_src->internal_md.type, __LINE__, __FILE__);
       *pstatus = CACHE_INODE_BAD_TYPE;
       pclient->stat.func_stats.nb_err_unrecover[CACHE_INODE_LINK] += 1;
       return *pstatus;
@@ -300,42 +300,42 @@ cache_inode_status_t cache_inode_link(cache_entry_t * pentry_src,
       V(pentry_src->lock);
 
       if (fsal_status.major == ERR_FSAL_STALE)
-	{
-	  cache_inode_status_t kill_status;
-	  fsal_status_t getattr_status;
+        {
+          cache_inode_status_t kill_status;
+          fsal_status_t getattr_status;
 
-	  DisplayLog
-	      ("cache_inode_link: Stale FSAL File Handle detected for at least one in  pentry = %p and pentry = %p",
-	       pentry_src, pentry_dir_dest);
+          DisplayLog
+              ("cache_inode_link: Stale FSAL File Handle detected for at least one in  pentry = %p and pentry = %p",
+               pentry_src, pentry_dir_dest);
 
-	  /* Use FSAL_getattrs to find which entry is staled */
-	  getattr_status = FSAL_getattrs(&handle_src, pcontext, &link_attributes);
-	  if (getattr_status.major == ERR_FSAL_ACCESS)
-	    {
-	      DisplayLog
-		  ("cache_inode_link: Stale FSAL File Handle detected for pentry = %p",
-		   pentry_src);
+          /* Use FSAL_getattrs to find which entry is staled */
+          getattr_status = FSAL_getattrs(&handle_src, pcontext, &link_attributes);
+          if (getattr_status.major == ERR_FSAL_ACCESS)
+            {
+              DisplayLog
+                  ("cache_inode_link: Stale FSAL File Handle detected for pentry = %p",
+                   pentry_src);
 
-	      if (cache_inode_kill_entry(pentry_src, ht, pclient, &kill_status) !=
-		  CACHE_INODE_SUCCESS)
-		DisplayLog("cache_inode_link: Could not kill entry %p, status = %u",
-			   pentry_src, kill_status);
-	    }
+              if (cache_inode_kill_entry(pentry_src, ht, pclient, &kill_status) !=
+                  CACHE_INODE_SUCCESS)
+                DisplayLog("cache_inode_link: Could not kill entry %p, status = %u",
+                           pentry_src, kill_status);
+            }
 
-	  getattr_status = FSAL_getattrs(&handle_dest, pcontext, &link_attributes);
-	  if (getattr_status.major == ERR_FSAL_ACCESS)
-	    {
-	      DisplayLog
-		  ("cache_inode_link: Stale FSAL File Handle detected for pentry = %p",
-		   pentry_dir_dest);
+          getattr_status = FSAL_getattrs(&handle_dest, pcontext, &link_attributes);
+          if (getattr_status.major == ERR_FSAL_ACCESS)
+            {
+              DisplayLog
+                  ("cache_inode_link: Stale FSAL File Handle detected for pentry = %p",
+                   pentry_dir_dest);
 
-	      if (cache_inode_kill_entry(pentry_dir_dest, ht, pclient, &kill_status) !=
-		  CACHE_INODE_SUCCESS)
-		DisplayLog("cache_inode_link: Could not kill entry %p, status = %u",
-			   pentry_dir_dest, kill_status);
-	    }
+              if (cache_inode_kill_entry(pentry_dir_dest, ht, pclient, &kill_status) !=
+                  CACHE_INODE_SUCCESS)
+                DisplayLog("cache_inode_link: Could not kill entry %p, status = %u",
+                           pentry_dir_dest, kill_status);
+            }
 
-	}
+        }
       *pstatus = CACHE_INODE_FSAL_ESTALE;
 
       return *pstatus;
@@ -344,8 +344,8 @@ cache_inode_status_t cache_inode_link(cache_entry_t * pentry_src,
   /* Post an asynchronous operation */
   P(pclient->pool_lock);
   GET_PREALLOC(pasyncopdesc,
-	       pclient->pool_async_op,
-	       pclient->nb_pre_async_op_desc, cache_inode_async_op_desc_t, next_alloc);
+               pclient->pool_async_op,
+               pclient->nb_pre_async_op_desc, cache_inode_async_op_desc_t, next_alloc);
   V(pclient->pool_lock);
 
   if (pasyncopdesc == NULL)
@@ -404,11 +404,11 @@ cache_inode_status_t cache_inode_link(cache_entry_t * pentry_src,
     case REGULAR_FILE:
       /* If object is a data cached regular file, keeps it mtime and size, STEP 2 */
       if (pentry_src->object.file.pentry_content != NULL)
-	{
-	  pentry_src->object.file.attributes.mtime = save_mtime;
-	  pentry_src->object.file.attributes.filesize = save_size;
-	  pentry_src->object.file.attributes.spaceused = save_spaceused;
-	}
+        {
+          pentry_src->object.file.attributes.mtime = save_mtime;
+          pentry_src->object.file.attributes.filesize = save_size;
+          pentry_src->object.file.attributes.spaceused = save_spaceused;
+        }
       pentry_src->object.file.attributes.numlinks += 1;
       pentry_src->object.file.attributes.ctime.seconds = pasyncopdesc->op_time.tv_sec;
       pentry_src->object.file.attributes.ctime.nseconds = pasyncopdesc->op_time.tv_usec;
@@ -419,7 +419,7 @@ cache_inode_status_t cache_inode_link(cache_entry_t * pentry_src,
       pentry_src->object.symlink.attributes.numlinks += 1;
       pentry_src->object.symlink.attributes.ctime.seconds = pasyncopdesc->op_time.tv_sec;
       pentry_src->object.symlink.attributes.ctime.nseconds =
-	  pasyncopdesc->op_time.tv_usec;
+          pasyncopdesc->op_time.tv_usec;
       *pattr = pentry_src->object.symlink.attributes;
       break;
 
@@ -429,26 +429,26 @@ cache_inode_status_t cache_inode_link(cache_entry_t * pentry_src,
     case FIFO_FILE:
       pentry_src->object.special_obj.attributes.numlinks += 1;
       pentry_src->object.special_obj.attributes.ctime.seconds =
-	  pasyncopdesc->op_time.tv_sec;
+          pasyncopdesc->op_time.tv_sec;
       pentry_src->object.special_obj.attributes.ctime.nseconds =
-	  pasyncopdesc->op_time.tv_usec;
+          pasyncopdesc->op_time.tv_usec;
       *pattr = pentry_src->object.special_obj.attributes;
       break;
 
     default:
       DisplayLogJdLevel(pclient->log_outputs, NIV_CRIT,
-			"WARNING: Major type incoherency line %d in file %s", __LINE__,
-			__FILE__);
+                        "WARNING: Major type incoherency line %d in file %s", __LINE__,
+                        __FILE__);
       break;
     }
 
   /* Add the new entry in the destination directory */
   if (cache_inode_add_cached_dirent(pentry_dir_dest,
-				    plink_name,
-				    pentry_src,
-				    NULL,
-				    ht,
-				    pclient, pcontext, &status) != CACHE_INODE_SUCCESS)
+                                    plink_name,
+                                    pentry_src,
+                                    NULL,
+                                    ht,
+                                    pclient, pcontext, &status) != CACHE_INODE_SUCCESS)
     {
       V(pentry_dir_dest->lock);
       V(pentry_src->lock);
